Add skipped test for potential failure case
If we encountered more than 9 of a given character in a subsequence, the
current implementation would fail, as shown in this test. This is
because when incrementing the count, I'm really just incrementing the
codepoint value (which conveniently works for counts < 10).

I think it's possible that you would never in fact encounter sequences
as described above, but worth confirming - perhaps that depends on the
starting value?