Add rough GenStage skeleton for typeahead
This adds a rough skeleton for GenStage to be used with our typeahead
search function.

This is currently split into 3 stages:

\### `QueryProducer`

`QueryProducer` is unusual for a producer, as the state it holds is just 1
word, not many. This is because I only care about the most recent
version of whatever has been input by the user, and so only store this.

Unsure how this will behave with latency atm.

\### `QueryProducerConsumer`

`QueryProducerConsumer` takes care of grabbing the state from the
producer, at a configurable rate. I've taken this implementation from an
example in the Elixir docs, and think it might be more complex than I
need.

\### `QueryConsumer`

This simply prints the latest event to the screen.