Update GenStage to be aware of socket
This updates the GenStage implementation to be aware of sockets, meaning
we can update the producer for a given channel, and push from the
consumer to the channel it knows about.

This is good in that it means we could have more than one user on this
at once, and they'd only see their own messages. However, the rate limit
will now be exceeded if the limit is per-application, rather than
per-socket, which seems likely.