Use internal GenStage event/demand buffers
This switches to using the internal buffers in `Query`. This handles
buffering somewhat better than my manual implementation, as for
instance:

* if we receive events but have not enough demand, we enqueue them. We
set this buffer size to 1, so that we never enqueue more than 1 event,
which makes sense for our purposes (no point rendering stale search
results).
* if receive demand but no events, the next events to come through are
dispatched immediately, and so the search can happen straight away
instead of at the next demand.

This is much better than what we had before. The main issue I see with
it is that as far as I can see there is no way to set the max pending
demand buffer, so this would seem to be unbounded. I'd rather limit this
to say 100 (could likely handle this somehow in the producer_consumer.